+++ velox/exec/benchmarks/CMakeLists.txt
@@ -40,6 +40,16 @@ target_link_libraries(
   Folly::follybenchmark
 )
 
+add_executable(velox_local_exchange_benchmark LocalExchangeBenchmark.cpp)
+
+target_link_libraries(
+  velox_local_exchange_benchmark
+  velox_exec
+  velox_exec_test_lib
+  velox_vector_test_lib
+  Folly::follybenchmark
+)
+
 add_executable(velox_merge_benchmark MergeBenchmark.cpp)
 
 target_link_libraries(
